move FirefoxOptions to new file src/capabilities.rs
This provides slightly better code organisation and reduces length of
the src/marionette.rs file.
